Remove legacy rosbuild files (#231)
This package is distributed as a standalone Python package in all
versions of ROS, and has been for many releases. This manifest and
CMakeLists.txt are from the transitional days around when catkin first
emerged, when it appears that rospkg was distributed as a ROS package
instead of a standalone python package.

The presence of a 'dry' manifest is confusing colcon and serves no
useful purpose anymore.