Avoid use of deprecated function in distro module.
Recent versions of the distro module have deprecated the
`distro.linux_distribution()` which is a compatibility shim for the
`platform.linux_distribution()` function removed in python 3.8.

All versions of the distro library going back to at least 1.0.1[1]
packaged in Ubuntu Bionic[2] provide the `id`, `version`, and `codename`
methods which are recommended instead.

`distro.id()` is used instead of `distro.name()` corresponding with the
full_distribution_name=0 argument to `linux_distribution`.

When updating the tests to support this change I noticed that the mocks
were being added in such a way that the `dist` conditional branch would never
actually be tested. I didn't actually look at how far back we'd have to
go to find a platform module which did not support `linux_distribution`
but it made sense to at least make testing that branch possible.

We'll have to take care that the `codename` member of the tuple
continues to match what our scripts expect downstream since the tests
mock the distro return values we don't have full coverage of the actual
results on contemporary platforms.

[1]: https://github.com/python-distro/distro/blob/v1.0.1/
[2]: https://packages.ubuntu.com/bionic/python3-distro