Add Conda OS detection. (#224)
We are currently adding rosdep functionality to the RoboStack project which builds ROS binaries with Conda. The first step towards this aim is adding an OsDetector here. Locally, we already have a working version of rosdep that makes use of the new RoboStack OsDetector, and would like to merge as many changes as possible upstream. See RoboStack/ros-noetic#41 for more information

This modification was inspired by #171. In that case, Nix could be used either as the package manager of the NixOs operating system, as a "user" package manager on any Linux distribution, a bit similar to how conda works.