Use FdoDetect to identify Raspbian systems (#257)
Raspbian Buster was close enough to vanilla Debian Buster that the
implementation of the Debian detector in os_detect identified it as
such. Things changed in Raspbian Bullsye, and it became necessary to
start identifying Raspbian as a discrete platform. This commit makes
that change.

For Raspbian Bullseye, this is not a breaking change because the
previous behavior was to fail to detect. On Raspbian Buster, this change
will break rosdep versions prior to 0.22.0, where the alias was added.

Another note is that the FdoDetect implementation doesn't output the
minor version number, but I'm not aware of any downstream code using the
version number of Debian platforms at all.